feat: "实现AccountController"

+++ src/main/java/com/javaee/elderlycanteen/entity/Account.java
@@ -28,7 +28,7 @@ public class Account {
     private String name;            // 姓名
     private String idCard;          // 身份证号码
     private String portrait;        // 头像
-    private BigDecimal money;       // 账户余额
+    private Double money;       // 账户余额
     private String gender;
 
 }
